requirements.txt
asgiref==3.7.2
astroid==2.15.6
Babel==2.12.1
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.7
Django==4.2.4
django-import-export==3.3.1
django-jazzmin==2.6.0
django-money==3.2.0
django-multiselectfield==0.1.12
django-postgres-extensions==0.9.3
django-simple-history==3.3.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
isort==5.12.0
lazy-object-proxy==1.9.0
MarkupPy==1.14
mccabe==0.7.0
numpy==1.25.2
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.3
platformdirs==3.10.0
psycopg2==2.9.7
py-moneyed==3.0
pylint==2.17.5
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tomlkit==0.12.1
typing_extensions==4.7.1
tzdata==2023.3
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0
